Move Welcome page to separate route
